backend/requirements.txt
# Core Framework
flask==3.0.0
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-limiter==3.5.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pydantic==2.5.0

# Security
python-dotenv==1.0.0
bcrypt==4.1.2

# Data Analysis
pandas==2.1.4
numpy==1.26.2

# OFX Import (PRIORITY)
ofxparse==0.21

# Machine Learning (PRIORITY)
scikit-learn==1.7.2
joblib==1.3.2

# AI / OpenAI Integration
openai==2.8.1

# Excel Reading (temporary - only for ML training)
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
black==23.12.1
ruff==0.1.9
mypy==1.7.1

# Production Server
gunicorn==21.2.0
